+++ src/com/cburch/logisim/verilog/comp/factories/wordlvl/MemoryOpFactory.java
@@ -1,9 +1,9 @@
 package com.cburch.logisim.verilog.comp.factories.wordlvl;
 
 import com.cburch.logisim.verilog.comp.AbstractVerilogCellFactory;
-import com.cburch.logisim.verilog.comp.VerilogCell;
+import com.cburch.logisim.verilog.comp.impl.VerilogCell;
 import com.cburch.logisim.verilog.comp.VerilogCellFactory;
-import com.cburch.logisim.verilog.comp.WordLvlCellImpl;
+import com.cburch.logisim.verilog.comp.impl.WordLvlCellImpl;
 import com.cburch.logisim.verilog.comp.auxiliary.CellType;
 import com.cburch.logisim.verilog.comp.specs.GenericCellAttribs;
 import com.cburch.logisim.verilog.comp.specs.wordlvl.MemoryOp;
