+++ src/com/cburch/logisim/verilog/comp/factories/wordlvl/UnaryOpFactory.java
@@ -1,8 +1,8 @@
 package com.cburch.logisim.verilog.comp.factories.wordlvl;
 
 import com.cburch.logisim.verilog.comp.AbstractVerilogCellFactory;
-import com.cburch.logisim.verilog.comp.VerilogCell;
-import com.cburch.logisim.verilog.comp.WordLvlCellImpl;
+import com.cburch.logisim.verilog.comp.impl.VerilogCell;
+import com.cburch.logisim.verilog.comp.impl.WordLvlCellImpl;
 import com.cburch.logisim.verilog.comp.auxiliary.CellType;
 import com.cburch.logisim.verilog.comp.specs.CommonOpAttribs;
 import com.cburch.logisim.verilog.comp.specs.wordlvl.UnaryOp;
