+++ src/com/cburch/logisim/verilog/comp/impl/AbstractVerilogCell.java
@@ -1,4 +1,4 @@
-package com.cburch.logisim.verilog.comp;
+package com.cburch.logisim.verilog.comp.impl;
 
 import com.cburch.logisim.verilog.comp.auxiliary.CellType;
 import com.cburch.logisim.verilog.comp.auxiliary.PortEndpoint;
