+++ src/com/cburch/logisim/verilog/comp/impl/GateLvlCell.java
@@ -1,4 +1,4 @@
-package com.cburch.logisim.verilog.comp;
+package com.cburch.logisim.verilog.comp.impl;
 
 public interface GateLvlCell extends VerilogCell {
 }
