+++ src/com/cburch/logisim/verilog/comp/impl/WordLvlCell.java
@@ -1,4 +1,4 @@
-package com.cburch.logisim.verilog.comp;
+package com.cburch.logisim.verilog.comp.impl;
 
 public interface WordLvlCell extends VerilogCell {
 }
