+++ src/com/cburch/logisim/verilog/std/wordlvl/BinaryOp.java
@@ -1,8 +1,5 @@
 package com.cburch.logisim.verilog.std.wordlvl;
 
-import com.cburch.logisim.verilog.comp.AbstractVerilogCell;
-import com.cburch.logisim.verilog.comp.GateLvlCell;
-
 public class BinaryOp {
 
 }
